Moved errors to errors.ts file for axios-hal-json package
