+++ packages/axios-hal-json/src/followHalJson/follow-hal-json.ts
@@ -1,53 +1,7 @@
 import axios, { AxiosRequestConfig, AxiosResponse } from "axios";
+import { HalJsonRequestChainError, NoHalJsonLinkToFollowError, NoHalJsonLinksInResponseError, NoHalJsonTemplateValueError } from "../errors";
 import "../index";
 
-/**
- * A request in the follow-chain failed
- * @category Error
- */
-export class HalJsonRequestChainError extends Error {
-  // eslint-disable-next-line no-unused-vars
-  constructor(public config: AxiosRequestConfig, public originalError: Error) {
-    super(`Request in hal-json chain failed for config: ${JSON.stringify(config)}`);
-    Object.setPrototypeOf(this, HalJsonRequestChainError.prototype);
-  }
-}
-
-/**
- * A response within the follow-chain does not contain _links.
- * @category Error
- */
-export class NoHalJsonLinksInResponseError extends Error {
-  // eslint-disable-next-line no-unused-vars
-  constructor(public config: AxiosRequestConfig, public response: AxiosResponse) {
-    super(`No _links found in response: ${JSON.stringify(response.data)}`);
-    Object.setPrototypeOf(this, NoHalJsonLinksInResponseError.prototype);
-  }
-}
-
-/**
- * A response in the follow-chain does not contain a follow link in _links.
- * @category Error
- */
-export class NoHalJsonLinkToFollowError extends Error {
-  // eslint-disable-next-line no-unused-vars
-  constructor(public follow: string, public config: AxiosRequestConfig, public response: AxiosResponse) {
-    super(`No href for '${follow}' found in _links: ${JSON.stringify(response.data._links)}`);
-    Object.setPrototypeOf(this, NoHalJsonLinkToFollowError.prototype);
-  }
-}
-
-/**
- * A link in the follow-chain needs to be templated but no template value was given.
- * @category Error
- */
-export class NoHalJsonTemplateValueError extends Error {
-  // eslint-disable-next-line no-unused-vars
-  constructor(public template: string, public followUrl: string, public config?: AxiosRequestConfig, public response?: AxiosResponse) {
-    super(`No template value for '${template}'. Needed for url: '${followUrl}`);
-    Object.setPrototypeOf(this, NoHalJsonTemplateValueError.prototype);
-  }
-}
 
 /**
  * This method can be registered as a [axios interceptor]{@link https://github.com/axios/axios#interceptors} to add hal-json follow behaviour. After registering just add ```follows``` and ```templates``` to your [request-config]{@link https://github.com/axios/axios#request-config}. To work with relative hal-json links use the ```baseUri``` and ```uri``` property.
